+++ Userland/Libraries/LibGUI/JsonArrayModel.cpp
@@ -59,7 +59,7 @@ bool JsonArrayModel::set(int row, Vector<JsonValue>&& values)
 {
     VERIFY(values.size() == m_fields.size());
 
-    if (row >= m_array.size())
+    if ((size_t)row >= m_array.size())
         return false;
 
     JsonObject obj;
@@ -76,12 +76,12 @@ bool JsonArrayModel::set(int row, Vector<JsonValue>&& values)
 
 bool JsonArrayModel::remove(int row)
 {
-    if (row >= m_array.size())
+    if ((size_t)row >= m_array.size())
         return false;
 
     JsonArray new_array;
-    for (int i = 0; i < m_array.size(); ++i)
-        if (i != row)
+    for (size_t i = 0; i < m_array.size(); ++i)
+        if (i != (size_t)row)
             new_array.append(m_array.at(i));
 
     m_array = new_array;
